requirements.txt
streamlit==1.52.0
google-cloud-vision==3.8.0
Pillow==12.0.0
requests==2.32.3
fpdf2==2.8.1
python-dotenv==1.0.1
pandas==2.3.3
numpy==2.3.5